backend/requirements.txt
streamlit
PyPDF2
scikit-learn==1.7.2
pandas
numpy
google-generativeai
python-dotenv
requests
xgboost
